[beef-tool] Create workflow job for beef-tool executable
